scripts Drop/Accept for IP tables run into issue due to ip address change. Explanation: IP ADDRESSES OF THE registry.agt.thousandeyes.com change from time to time. Previous script did dig and stored output to address variable.
When this change is done while TSing or machine was powered off and address is changed, above script will fail to remove iptables rules as it is not in table - it is doing dig and get new IP address! Hence, new script that will store ip addresses to a file, read from a file. I will add more improvements to check if rule exist before trying to remove.