Changed db logic Added to return task status to all GET pages mater task id = 0 if master task is 1 all tasks should be enabled. if any task status changes to 0 it should also change master task to 0.
If master task status is 1 all task should be enabled. To ensure that, we will attempt to add total_number of tasks to db. Then we will run DNS fix, as we depend on DNS to get list of ip addresses. Once that is done we will break all tasks. DNS break MUST be last with delay always.

If master task is not 1, lab was already used, so we leave tasks untouched.